previous_requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
backports.tarfile==1.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.5.1
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.8.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.0
onnxruntime==1.16.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.1.1
parso==0.8.4
pefile==2023.2.7
Pillow==10.0.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==6.30.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.4
scikit-image==0.22.0
scipy==1.11.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
tifffile==2025.3.13
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
